Added known error code when shutdown is already complete
