+++ process.go
@@ -2,8 +2,6 @@ package hcsshim
 
 import (
 	"encoding/json"
-	"errors"
-	"fmt"
 	"io"
 	"syscall"
 	"time"
@@ -11,13 +9,7 @@ import (
 	"github.com/Sirupsen/logrus"
 )
 
-type ProcessError struct {
-	Process   *process
-	Operation string
-	ExtraInfo string
-	Err       error
-}
-
+// ContainerError is an error encountered in HCS
 type process struct {
 	handle         hcsProcess
 	processID      int
@@ -48,7 +40,7 @@ type closeHandle struct {
 }
 
 type processStatus struct {
-	ProcessId      uint32
+	ProcessID      uint32
 	Exited         bool
 	ExitCode       uint32
 	LastWaitResult int32
@@ -223,7 +215,7 @@ func (process *process) properties() (*processStatus, error) {
 	}
 
 	if propertiesp == nil {
-		return nil, errors.New("Unexpected result from hcsGetProcessProperties, properties should never be nil")
+		return nil, ErrUnexpectedValue
 	}
 	propertiesRaw := convertAndFreeCoTaskMemBytes(propertiesp)
 
@@ -395,47 +387,4 @@ func (process *process) unregisterCallback() error {
 	handle = 0
 
 	return nil
-}
-
-func (e *ProcessError) Error() string {
-	if e == nil {
-		return "<nil>"
-	}
-
-	if e.Process == nil {
-		return "Unexpected nil process for error: " + e.Err.Error()
-	}
-
-	s := fmt.Sprintf("process %d", e.Process.processID)
-
-	if e.Process.container != nil {
-		s += " in container " + e.Process.container.id
-	}
-
-	if e.Operation != "" {
-		s += " " + e.Operation
-	}
-
-	if e.Err != nil {
-		s += fmt.Sprintf(" failed in Win32: %s (0x%x)", e.Err, win32FromError(e.Err))
-	}
-
-	return s
-}
-
-func makeProcessError(process *process, operation string, extraInfo string, err error) error {
-	// Don't wrap errors created in hcsshim
-	if err == ErrTimeout ||
-		err == ErrUnexpectedProcessAbort ||
-		err == ErrUnexpectedContainerExit ||
-		err == ErrHandleClose ||
-		err == ErrInvalidProcessState ||
-		err == ErrInvalidNotificationType ||
-		err == ErrVmcomputeOperationPending {
-		return err
-	}
-
-	processError := &ProcessError{Process: process, Operation: operation, ExtraInfo: extraInfo, Err: err}
-	logrus.Error(processError)
-	return processError
 }
